Commit(feature/9/Implemented-Koin):
-Se integra Koin para la inyección de dependencias.
-Se reestructura y reorganizan los paquetes del proyecto.
-Se limpia el archivo .gitignore.
