+++ app/src/main/java/com/example/simpsons/SimpsonsApplication.kt
@@ -1,0 +1,20 @@
+package com.example.simpsons
+
+import android.app.Application
+import com.example.simpsons.core.di.AppModule
+import org.koin.android.ext.koin.androidContext
+import org.koin.android.ext.koin.androidLogger
+import org.koin.core.context.startKoin
+import org.koin.ksp.generated.module
+
+class SimpsonsApplication : Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+        startKoin {
+            androidLogger()
+            androidContext(this@SimpsonsApplication)
+            modules(AppModule().module)
+        }
+    }
+}
