+++ app/src/main/java/com/example/simpsons/core/di/AppModule.kt
@@ -1,0 +1,8 @@
+package com.example.simpsons.core.di
+
+import org.koin.core.annotation.ComponentScan
+import org.koin.core.annotation.Module
+
+@Module
+@ComponentScan("com.example.simpsons")
+class AppModule
