+++ app/src/main/java/com/example/simpsons/core/errors/ErrorApp.kt
@@ -1,0 +1,9 @@
+package com.example.simpsons.core.errors
+
+sealed class ErrorApp: Throwable() {
+
+    object InternetConexionError : ErrorApp()
+
+    object ServerErrorApp : ErrorApp()
+
+}
