+++ app/src/main/java/com/example/simpsons/features/data/SimpsonsDataRepository.kt
@@ -1,0 +1,25 @@
+package com.example.simpsons.features.data
+
+import com.example.simpsons.features.data.remote.api.SimpsonsApiRemoteDataSource
+import com.example.simpsons.features.data.remote.api.toModel
+import com.example.simpsons.features.domain.Simpson
+import com.example.simpsons.features.domain.SimpsonsRepository
+import org.koin.core.annotation.Single
+
+@Single
+class SimpsonsDataRepository(
+    private val apiRemoteDataSource: SimpsonsApiRemoteDataSource
+) : SimpsonsRepository {
+
+    override suspend fun findAll(): Result<List<Simpson>> {
+        return apiRemoteDataSource.getSimpsons().map { apiModelsList ->
+            apiModelsList.map { apiModel -> apiModel.toModel() }
+        }
+    }
+
+    override suspend fun findById(id: String): Result<Simpson> {
+        return apiRemoteDataSource.getSimpsonById(id).map { apiModel ->
+            apiModel.toModel()
+        }
+    }
+}
