+++ app/src/main/java/com/example/simpsons/features/data/remote/api/SimpsonsApiResponse.kt
@@ -1,0 +1,5 @@
+package com.example.simpsons.features.data.remote.api
+
+data class SimpsonsApiResponse(
+    val results: List<SimpsonsApiModel>
+)
