+++ app/src/main/java/com/example/simpsons/features/domain/Simpson.kt
@@ -1,0 +1,8 @@
+package com.example.simpsons.features.domain
+
+data class Simpson(
+    val id: String,
+    val name: String,
+    val phrase: String,
+    val imageUrl: String
+)
