+++ app/src/main/java/com/example/simpsons/features/domain/SimpsonRepository.kt
@@ -1,0 +1,6 @@
+package com.example.simpsons.features.domain
+
+interface SimpsonsRepository {
+    suspend fun findAll(): Result<List<Simpson>>
+    suspend fun findById(id: String): Result<Simpson>
+}
