--- app/src/main/java/com/example/simpsons/features/simpsons/core/api/ApiClient.kt
@@ -1,17 +1,0 @@
-package com.example.simpsons.features.simpsons.core.api
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class ApiClient {
-    private val BASE_URL = "https://thesimpsonsapi.com/"
-
-    private val retrofit = Retrofit.Builder()
-        .baseUrl(BASE_URL)
-        .addConverterFactory(GsonConverterFactory.create())
-        .build()
-
-    fun <T> createService(typeClass: Class<T>): T {
-        return retrofit.create(typeClass)
-    }
-}
