--- app/src/main/java/com/example/simpsons/features/simpsons/data/remote/api/SimpsonsApiModel.kt
@@ -1,13 +1,0 @@
-package com.example.simpsons.features.simpsons.data.remote.api
-
-import com.google.gson.annotations.SerializedName
-
-data class SimpsonsApiModel(
-    val id: Int,
-    val name: String,
-
-    @SerializedName("portrait_path")
-    val portraitPath: String,
-
-    val phrases: List<String>
-)
