--- app/src/main/java/com/example/simpsons/features/simpsons/data/remote/api/SimpsonsApiResponse.kt
@@ -1,5 +1,0 @@
-package com.example.simpsons.features.simpsons.data.remote.api
-
-data class SimpsonsApiResponse(
-    val results: List<SimpsonsApiModel>
-)
