--- app/src/main/java/com/example/simpsons/features/simpsons/domain/ErrorApp.kt
@@ -1,9 +1,0 @@
-package com.example.simpsons.features.simpsons.domain
-
-sealed class ErrorApp: Throwable() {
-
-    object InternetConexionError : ErrorApp()
-
-    object ServerErrorApp : ErrorApp()
-
-}
