--- app/src/main/java/com/example/simpsons/features/simpsons/domain/FetchSimpsonsUseCase.kt
@@ -1,7 +1,0 @@
-package com.example.simpsons.features.simpsons.domain
-
-class FetchSimpsonsUseCase(private val simpsonsRepository: SimpsonsRepository) {
-    suspend operator fun invoke(): Result<List<Simpson>> {
-        return simpsonsRepository.findAll()
-    }
-}
