--- app/src/main/java/com/example/simpsons/features/simpsons/domain/Simpson.kt
@@ -1,8 +1,0 @@
-package com.example.simpsons.features.simpsons.domain
-
-data class Simpson(
-    val id: String,
-    val name: String,
-    val phrase: String,
-    val imageUrl: String
-)
