--- app/src/main/java/com/example/simpsons/features/simpsons/domain/SimpsonRepository.kt
@@ -1,6 +1,0 @@
-package com.example.simpsons.features.simpsons.domain
-
-interface SimpsonsRepository {
-    suspend fun findAll(): Result<List<Simpson>>
-    suspend fun findById(id: String): Result<Simpson>
-}
