+++ app/src/main/res/layout/fragment_initialization.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/simpsons_background_placeholder"
-    tools:context="com.example.simpsons.features.simpsons.presentation.initialization.InitializationFragment">
+    tools:context="com.example.simpsons.features.presentation.initialization.InitializationFragment">
 
     <TextView
         android:id="@+id/titleTextView"
